Migrated 'align_images_using_mmi' and associated test into their own modules
